big update

+++ src/NodeData.py
@@ -65,8 +65,8 @@ class NodeData:
         if self.position is None or node.position is None:
             return -1
         else:
-            return np.sqrt((np.power(self.position[0] - node.position[0], 2)) +
-                           (np.power(self.position[1] - node.position[1], 2)))
+            return float(np.sqrt((np.power(self.position[0] - node.position[0], 2)) +
+                                 (np.power(self.position[1] - node.position[1], 2))))
 
     def set_scc(self, scc_key: int):
         """
